Create & link boilerplate recipe pages to homepage
The file directory for recipes is set up as-is so that when one clicks
a recipe from the homepage, that it opens up in a new tab and does not
mention the html document in the URL for clarity.

The anchor tag attribute rel is set to noopener and noreferrer for the
sake of security against tabnabbing.