Write the steak burrito bowl recipe
I do not claim to be the creator of this recipe, nor the one who made
this recipe available on the internet first. I copied the contents of
the recipe from Ethan Chlebowski.

Again, this project is meant to be for non-commercial & private learning
purposes of creating a basic recipes website with HTML. That's all.