Fix the external CSS href for recipe pages
Before the change, the styles.css was only being applied to the homepage
and not the recipe pages - even though it was intended to be applied to
all pages on the website.

With this change, the styles.css properly applies to both the homepage
and the recipe pages. I solved it by updating the relative links rather
than creating a dedicated styles.css for each of the recipe pages, for
the sake of simplicity.