Set size & responsive attributes for hero banners
Before this change, there was no specified image size HTML attributes,
nor way to specify the images that were hero banners in CSS. What was
wrong with this was it meant the browser layout would jump and flash
once the images loaded in, and that when they did load in, that they
clipped outside of the visible area when viewed in a split page mode.

Now, the images have specified height & width values that match their
source dimensions in HTML, and dedicated hero-banner class to help them
be more responsive to the width of the page they are on. I did it the
way I did, so that if there are images in the future, that they are able
to be distinguished from hero banner images.