Create navigational header between all pages
Before, there was no navigational header, which meant that once you
clicked onto a recipe, there was no way to go from the that tab back to
the homepage, or to quickly jump to viewing other recipes.

Now, if you click on the logo, you get sent back to the homepage, and if
you click on the recipes, you get sent directly to the all recipes
section of the homepage. I choose to do it this way, because I felt that
the homepage would be the ideal place to store all recipes, but if I
ever wanted to add content such as a featured recipe, to ensure that the
user can jump straight to the relevant section with minimal scrolling.