Revamp the hero banners for recipe pages
Before, the hero banner background image was not responsive to screen
changes at all, and the H1 was very far from the rest of the content,
which just looked ugly. And there was no proper credit given to who
I yoinked the recipe from on the pages either.

Now, the hero banner is somewhat responsive to screen size, and the H1
and the source of the recipe is overlaid over the hero banner, in a
more aesthetic way.

I also decided to put a main tag in proactively, because the hero
banners on each recipe page will be the only content to spread across
the entirety of the view width. The rest will have dedicated padding,
eventually, to be more centered in the screen.