Setup margin & padding for recipes on desktop
Before, there wasn't any custom margin or padding on the recipe pages.
What was wrong: the readability was TERRIBLE.

The way it is now is a step-up, where the readability is acceptable and
the margins & padding are more conducive to a typical flow of reading.

Side effects of this change is that this was set up for a full screen
desktop viewport - it is NOT responsive to viewport widths, which means
the readability will not be guaranteed for other sizes.

I still did it like this, because of a lack of technical knowledge.
When I know how to do responsive design later in The Odin Project, I'll
hopefully revisit and revamp it then.