Revamp ingredients & steps lists to content cards
Before this change, they were both just walls of texts, and was ugly.
After this change, they're visually more distinct from each other,
which was the overall goal of this revamp.

Side-effect of this change is that the list-style-position property
default of 'outside' led to the list markers for <ul> and <ol> not
being visible due to margin / padding being changed from the default
values for <li>, so I had to change it to 'inside' for the <ul>, and
just added numbers to the <h3> elements in the <ol> list items, because
I could not get the list markers to line up cleanly with the <h3>
elements for the <ol> otherwise, and the custom set margin and padding
from prior commits were fine in all other considerations.