Adjust the target of the yoinked recipes content
Before, the id aka target of the yoinked recipes was the <ul> not the
<section> itself, which was fine then.

Now, the id was adjusted to be assigned to the <section> of yoinked
recipes, in preparation for doing a new navigational link feature.
Specifically, I wanted to establish the precedent that all navigational
links will typically jump to a id associated to a <section> element.