Stop opening new tabs & showing index.html files
Before, it was a designed behavior to open up a new tab when accessing
a recipe page, because I did not have a way to jump back to the home
page otherwise. Now, it's just a nuisance that lost its use case,
and it was been adjusted back to default <a> behavior.

Additionally, when jumping back to the home page from a recipe page, it
was observed that the index.html would be showing in the URL. This was
NOT intended behavior, and was adjusted to not reveal it.