Adjust footer padding & positioning
Before the footer was too thin and crunched up next to the rest of the
content on a page. Now, it matches the header padding, and to maintain
its past positioning, updated it using flex.

Additional side effect, to ensure that on all pages the footer would be
at the bottom, with no content or empty space below it, I adjusted the
min-height of the <body> accordingly.