Add navigational links for recipe pages
Before, if a user wanted to get to a specific section, they were forced
to scroll, which was annoying. Now, the user can just click on the nav
featured on each recipe page to just to its respective description,
ingredients, and steps.

Side effects include: issues of typography, lack of responsive design,
the coloring probably not being accessible or compliant, and there is
no option to jump back to the top, etc.

But side effects like typography and colors will be addressable when I
learn about how to adjust such for improved readability & accessibility.
Likewise, side effects like responsive design, jump back to top, sticky
headers, etc - will be addressable when I engage the JavaScript portion
of The Odin Project.