Fix the invalid HTML for viewing featured recipe
Before, there was a <button> element wrapped in an <a> element in order
to make an HTML button behave like a link. This approach was considered
invalid HTML when running it in a HTML validator, and thus I fixed the
HTML to be valid, by styling the <a> like a button instead.

There does exist a debate why on doing this is bad or neutral, but I
wanted this intended effect, simple as.