Set the meta viewport to look good on all devices
Unsurprising consequence - in the absence of responsive design or an
intentionally designed website for both mobile and desktop, this will
actually just make the website straight up unusable for mobile viewing.

Why change it? Lighthouse audit insists to have this added to ensure
pages are mobile friendly - which is true, in general circumstances.

I decided to do this now, so if I ever view this website in mobile, that
I remember why responsive design is important to do.