Optimize the image loading performance across site
Before, I was using non-compressed, non-resized, not defined width nor
height images, which got marked as diagnostic guidance in Google's
Lighthouse audit.

Now, with compressed, resized, and defined images, page performance is
better across the board, and Google's Lighthouse audit doesn't get mad.

Side-effects: Because the thumbnail images are specifically sized for
the current size of the thumbnail in the yoinked recipes section of the
home page, that might be an obstacle for a responsive design overhaul.

In case I forget what I did here specifically:
- Created dedicated re-sized & compressed thumbnails for the all recipes
section of the homepage
- Compressed images for the hero banners
- Set up the default image format to be webp, and fallbacks to jpg
- Set up preloads for the largest content picture all of the pages
- Ensured images have defined width & height within HTML