Add citations & disclaimer to all recipe pages
Before, there was no citations for the recipes, which would be blatant
content theft, which is very cringe to do. It is not enough to simply
acknowledge that the content is not my original work in comments like
these, as I did before.

Now, there are citations for each recipe at the bottom of each recipe
page, styled in MLA 9th Edition using Scribbr's free citation generator.
I ensured the styling would be as similar to the MLA style where it
ought to be, and also made sure to properly protect all the links from
tabnabbing that led to the respective sources.