Improve the color contrast ratio of text content
Before, the grey accent color that was something used as a background
color for text content would not be in compliance with WCAG contrast
ratio standards, I used https://webaim.org/resources/contrastchecker/
to ensure that the new colors are in full compliance.

Adjusted the darker grey accent color for the nav to a dark purple to
be more cohesive with the purple accent color.